Warning alerts added in the UI. dll index created for service_filetypes on filetype

+++ analyze/src/main/java/com/sc/fe/analyze/service/FiletypeExtensionsService.java
@@ -85,7 +85,9 @@ public class FiletypeExtensionsService {
         filetypeExtensionsRepo.delete(filetypeObj);
 
         ServiceFiletypes serviceFiletype = serviceFiletypeRepo.findByKeyFiletype(file_type);
-        serviceFiletypeRepo.delete(serviceFiletype);
+        if (serviceFiletype != null) {
+            serviceFiletypeRepo.delete(serviceFiletype);
+        }
 
         cacheService.evictAllCacheValues("ExtnFileMap");
     }
